rename peakgroup -> selection
