+++ alphadia/search/selection/utils.py
@@ -2,7 +2,7 @@ import numba as nb
 import numpy as np
 from numba.extending import overload
 
-from alphadia.search.peakgroup.fft import NumbaContextOnly
+from alphadia.search.selection.fft import NumbaContextOnly
 from alphadia.utilities.fragment_container import FragmentContainer
 from alphadia.utils import USE_NUMBA_CACHING
