+++ alphadia/workflow/peptidecentric/extraction_handler.py
@@ -9,8 +9,8 @@ from alphadia.plexscoring.config import CandidateConfig
 from alphadia.plexscoring.plexscoring import CandidateScoring
 
 # TODO: these imports could be conditional: HybridCandidateConfig, HybridCandidateSelection, CandidateConfig, CandidateScoring
-from alphadia.search.peakgroup.config_df import HybridCandidateConfig
-from alphadia.search.peakgroup.search import HybridCandidateSelection
+from alphadia.search.selection.config_df import HybridCandidateConfig
+from alphadia.search.selection.search import HybridCandidateSelection
 
 try:
     from alphadia_ng import PeakGroupSelection, SelectionParameters
